requirements.txt
pymongo
python-dotenv
redis